moving around some code and classes
