+++ java/README.md
@@ -1,0 +1,6 @@
+Welcome To The New Evolution In Sauce Labs Testing.
+The goal of Simple Sauce is to make your test automation and integration in Sauce Labs, simple :)
+
+Start a session on latest Mac/Safari:
+
+```webDriver = new SauceSession().withSafari().start();```
