+++ java/src/main/java/com/saucelabs/simplesauce/SafariVersion.java
